lpt_drv: fix gLptHz gLastError definitions
Declare these variables as globals, not static, as they are
referenced in another file.

Fix extern definition of gLastError to use the correct type (it is
defined unsigned short not unsigned long).

+++ lpt_drv/ksbiglptmain.c
@@ -27,7 +27,7 @@
 #include "ksbiglptmain.h"   
 
 extern unsigned long gLptHz;
-extern unsigned long gLastError;
+extern unsigned short gLastError;
 //========================================================================
 module_init(KModInit);
 module_exit(KModExit);
